This is a major change for the piping of DOM nodes.
Events can now be scoped. Events emitted from an emitter will also emit an additional event with the event name prefixed with 'EMITTER.'. If the the event originates from a DOM node it is prefixed with 'DOM.'. This allows you to bind events to piped emitters without overloading the DOM with event handers. You can also make sure an event is not spoofed and only originated from the DOM. If a user attempts to fire 'DOM.myEvent' an exemption will be thrown.

Only one hander for each event will be bound to each DOM node. Any additional handers will be called by the existing hander.